made a styled component for dialogcontenttext
