+++ src/dndApp/FeatOrTrait.tsx
@@ -1,6 +1,7 @@
-import { Typography, DialogTitle, DialogContent, DialogContentText, Stack, DialogActions, Button, useTheme } from "@mui/material";
+import { Typography, DialogTitle, DialogContent, DialogActions, Button } from "@mui/material";
 import { useState } from "react";
 import { StyledDialog } from "./StyledDialog";
+import { StyledDialogContentText } from "./StyledDialogContentText";
 
 interface FeatOrTraitProps {
     name: string;
@@ -8,7 +9,6 @@ interface FeatOrTraitProps {
 }
 export function FeatOrTrait({ name, description }: FeatOrTraitProps) {
     const [open, setOpen] = useState(false)
-    const theme = useTheme();
 
     return (
         <>
@@ -17,19 +17,7 @@ export function FeatOrTrait({ name, description }: FeatOrTraitProps) {
                 <StyledDialog open={open} onClose={() => setOpen(false)}>
                     <DialogTitle>{name}</DialogTitle>
                     <DialogContent>
-                        <DialogContentText>
-                            <Stack spacing={1}>
-                                {description.map(d => {
-                                    return (
-                                        <Typography sx={{
-                                            color: theme.dialogContentTextTypography,
-                                        }}>
-                                            {d}
-                                        </Typography>
-                                    );     
-                                })}
-                            </Stack>
-                        </DialogContentText>
+                        <StyledDialogContentText text={description}/>
                     </DialogContent>
                     <DialogActions>
                         <Button onClick={() => setOpen(false)}>Close</Button>
