+++ src/dndApp/Skills.tsx
@@ -1,11 +1,12 @@
 // eslint-disable-next-line @typescript-eslint/consistent-type-imports
-import { Button, DialogActions, DialogContent, DialogContentText, DialogTitle, Stack, Typography, useTheme } from "@mui/material";
+import { Button, DialogActions, DialogContent, DialogTitle, Stack, Typography } from "@mui/material";
 import characterSheet from "./formatted-sheet.json";
 import { useGetSkillsDescriptionQuery } from "./dndApiSlice";
 import { useState } from "react";
 import { StyledStack } from "./StyledStack";
 import { StyledPaper } from "./StyledPaper";
 import { StyledDialog } from "./StyledDialog";
+import { StyledDialogContentText } from "./StyledDialogContentText";
 
 export function Skills() {
     const skills = characterSheet.skills;
@@ -34,7 +35,6 @@ export function Skill({ name, statValue }: SkillProps) {
     const [open, setOpen] = useState(false);
     const { data } = useGetSkillsDescriptionQuery(name);
     const skillName = `${name.slice(0,1).toUpperCase()}${name.slice(1)}`.replace(/-/g, " ")
-    const theme = useTheme();
 
     return (
         <>
@@ -59,13 +59,7 @@ export function Skill({ name, statValue }: SkillProps) {
             >
                 <DialogTitle>{skillName}</DialogTitle>
                 <DialogContent>
-                    <DialogContentText>
-                        <Typography sx={{
-                            color: theme.dialogContentTextTypography,
-                        }}>
-                            {data?.desc}
-                        </Typography>
-                    </DialogContentText>
+                    <StyledDialogContentText text={data?.desc} />
                 </DialogContent>
                 <DialogActions>
                     <Button onClick={() => setOpen(false)}>Close</Button>
